feat: Restructure the folder tree
